+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.2.0] - 2022-05-18
 ### Added
 - handle optin cookie
 - persist allocations in cookie, we want it stable even if config changes
@@ -25,5 +26,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - injectable log delegate
   - injectable http fetch delegate
 
-[Unreleased]: https://github.com/SymplifyConversion/sst-sdk-nodejs/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/SymplifyConversion/sst-sdk-nodejs/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/SymplifyConversion/sst-sdk-nodejs/releases/tag/v0.2.0
 [0.1.0]: https://github.com/SymplifyConversion/sst-sdk-nodejs/releases/tag/v0.1.0